requirements.txt
flask==3.1.0
gspread==6.1.4
google-auth==2.29.0
python-dotenv==1.0.1
bcrypt==4.2.0
weasyprint==62.3
gunicorn==22.0.0
